requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
pandas-ta>=0.3.14b
python-telegram-bot>=20.0
pyyaml>=6.0
schedule>=1.2.0
requests>=2.31.0
